scripts/load_to_bigquery.py
#!/usr/bin/env python3
"""
Script to load processed data into BigQuery.
"""
import os
import sys
import logging
from google.cloud import bigquery
from google.cloud import storage

# Setup logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('logs/bigquery_loading.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

def main():
    """Main BigQuery loading function"""
    logger.info("Starting BigQuery loading process")
    # Implementation will go here
    
    logger.info("Loading complete")

if __name__ == "__main__":
    # Create logs directory if it doesn't exist
    os.makedirs('logs', exist_ok=True)
    main()
