requirements.txt
# Core data processing
pandas>=1.3.0
numpy>=1.20.0
tqdm>=4.62.0

# Google Cloud
google-cloud-bigquery>=2.30.0
google-cloud-storage>=1.42.0

# Data visualization for exploration
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Data validation
great-expectations>=0.14.0

# Utils
PyYAML>=6.0
python-dotenv>=0.19.0

# Future use when adding dbt (commented out)
# dbt-core>=1.0.0
# dbt-bigquery>=1.0.0

# Future use when adding ML models (commented out)
# scikit-learn>=1.0.0
# tensorflow>=2.8.0
# xgboost>=1.5.0